+++ server/controllers/account.js
@@ -1,82 +1,74 @@
 const db = require('../dao').model("accounts");
 const dbError = require("../dao").dbError;
 const utils = require('../utils');
+const accountUtils = require('./utils/accountUtils');
 
 module.exports = {
-    list: list,
-    get: get,
-    edit: edit,
-    create: create,
+  list: list,
+  get: get,
+  edit: edit,
+  create: create,
 };
 
 function list(req, res, next){
-    let allAccounts = db.getAll();
-    Object.keys(allAccounts).forEach(id => {
-      let account= allAccounts[id];
-      account.alias = [];
-      if (!utils.isEmpty(account.title)){
-        account.alias.push(utils.toPinyin(account.title));
-      }
-      if (!utils.isEmpty(account.username)){
-        account.alias.push(utils.toPinyin(account.username))
-      }
-      if (!utils.isEmpty(account.password)){
-        account.alias.push(utils.toPinyin(account.password))
-      }
-    });
-    res.status = 200;
-    res.json({
-        success: true,
-        accounts: allAccounts
-    });
-    return;
+  let options = {};
+  let allAccounts = accountUtils.list(options);
+  res.status = 200;
+  res.json({
+    success: true,
+    accounts: allAccounts
+  });
+  return;
 }
 
 function get(req, res, next){
-    let uid = req.params.uid;
-    let account = db.getOne(uid);
-    if (account instanceof dbError){
-        res.status(account.statusCode);
-        res.json(account);
-    }else{
-        res.status(200);
-        res.json({
-            success: true,
-            account: account
-        });
-    }
-    return;
+  let options = {}
+  options.uid = req.params.uid;
+  let account = accountUtils.get(options);
+  if (account instanceof dbError){
+    res.status(account.statusCode);
+    res.json(account);
+  }else{
+    res.status(200);
+    res.json({
+      success: true,
+      account: account
+    });
+  }
+  return;
 }
 
 function edit(req, res, next){
-    let uid = req.params.uid;
-    let content = req.body.content;
-    let err = db.updateOne(uid, content);
-    if (err && err instanceof dbError){
-        res.status(err.statusCode);
-        res.json(err);
-    }else{
-        res.status(200);
-        res.json({
-            success: true,
-            message: "Successfully updated."
-        })
-    }
-    return;
+  let options = {};
+  options.uid = req.params.uid;
+  options.content = req.body.content;
+  let err = accountUtils.edit(options);
+  if (err && err instanceof dbError){
+    res.status(err.statusCode);
+    res.json(err);
+  }else{
+    res.status(200);
+    res.json({
+        success: true,
+        message: "Successfully updated."
+    })
+  }
+  return;
 }
 
 function create(req, res, next){
-    let content = req.body.content;
-    let uid = db.create(content);
-    if (uid instanceof dbError){
-        res.status(uid.statusCode);
-        res.json(uid);
-    }else{
-        res.status(200);
-        res.json({
-            success: true,
-            new_uid: uid
-        });
-    }
-    return;
+  let options = {};
+  options.content = req.body.content;
+  let uid = accountUtils.create(options);
+  if (uid instanceof dbError){
+    res.status(uid.statusCode);
+    res.json(uid);
+  }else{
+    res.status(200);
+    res.json({
+      success: true,
+      new_uid: uid
+    });
+  }
+  return;
 }
